[flang][NFC] Extract ArraySectionAnalyzer from OptimizedBufferization
Extract `ArraySectionAnalyzer` from `OptimizedBufferization` into a standalone
analysis utility so it can be reused by other passes (e.g., `ScheduleOrderedAssignments`).

This is an NFC change that moves the `ArraySectionAnalyzer` class and its helper
methods to `flang/Optimizer/Analysis/ArraySectionAnalyzer.h` and `.cpp`.

Also extracts the logic to detect if a designate is using the indices
of an elemental operation in storage order.
